split out db wrapper stuff, but code will NOT run atm
tests of just database_context.py are (mostly) passing
   - is_rem/is_pem I'm not sure, are these based on current JP rotation?
   - I fixed get_monster_evo_gem
   - get_evolution_tree_ids does not include the base ID
   - I didn't test get_all_events or get_dungeon_by_id

however absolutely definitely this won't actually run since MonsterIndex
requires stuff that's no longer in the database manager class. I briefly
tried to make that work by further splitting up the files but that
didn't go well so I just gave up so here's what I did so far hf
